+++ src/App.js
@@ -12,8 +12,8 @@ import Footer from "./components/Footer/Footer";
 require ("./App.css");
 
 function App() {
-  let [endcaps, setEndcaps] = useState([])
-  let [columnOrder,setColumnOrder] = useState([])
+  // let [endcaps, setEndcaps] = useState([])
+  // let [columnOrder,setColumnOrder] = useState([])
   let [updated, setUpdated] = useState(false)
 
   // state = {
@@ -94,118 +94,117 @@ function App() {
   //   })
   // }
 
-  const handleFetchEndcaps = () => {
-    fetch("http://localhost:5000/endcaps")
-    .then((response) => response.json())
-    .then((jsonData) => {
-      const allEndcaps = jsonData.allEndcaps;
-      setEndcaps(allEndcaps);
-      console.log(jsonData)
-    })
+  // const handleFetchEndcaps = () => {
+  //   fetch("http://localhost:5000/endcaps")
+  //   .then((response) => response.json())
+  //   .then((jsonData) => {
+  //     const allEndcaps = jsonData.allEndcaps;
+  //     return allEndcaps;
+  //     // setEndcaps(allEndcaps);
+  //   })
+  //   //   const endcapData = jsonData.allEndcaps;
+  //   //   fetch("http://localhost:5000/settings", {
+  //   //     // credentials: "include"
+  //   //   })
+  //   //   .then((response) => response.json())
+  //   //   .then((jsonData => {
+  //   //     const currentColNew = jsonData.settings.find((setting) => {
+  //   //       return setting.user === this.state.user._id 
+  //   //     })
+  //   //     const currentColumnOrder = currentColNew.columnOrder.endcapIds;
+  //   //     const newState = {
+  //   //       ...this.state,
+  //   //       endcaps: endcapData,
+  //   //       columns: {
+  //   //         ...this.state.columns,
+  //   //         "column-1": {
+  //   //           ...this.state.columns["column-1"],
+  //   //           endcapIds: currentColumnOrder,
+  //   //         }
+  //   //       },
+  //   //     }
+  //   //     this.setState(newState);
+  //   //   })
+  //   // )})
+  //     .catch((err) => console.log(err));
+  // }
   
-    //   const endcapData = jsonData.allEndcaps;
-    //   fetch("http://localhost:5000/settings", {
-    //     // credentials: "include"
-    //   })
-    //   .then((response) => response.json())
-    //   .then((jsonData => {
-    //     const currentColNew = jsonData.settings.find((setting) => {
-    //       return setting.user === this.state.user._id 
-    //     })
-    //     const currentColumnOrder = currentColNew.columnOrder.endcapIds;
-    //     const newState = {
-    //       ...this.state,
-    //       endcaps: endcapData,
-    //       columns: {
-    //         ...this.state.columns,
-    //         "column-1": {
-    //           ...this.state.columns["column-1"],
-    //           endcapIds: currentColumnOrder,
-    //         }
-    //       },
-    //     }
-    //     this.setState(newState);
-    //   })
-    // )})
-      .catch((err) => console.log(err));
-  }
-  
-  useEffect(() => {
-    handleFetchEndcaps();
-    if(updated){
-      fetch("http://localhost:5000/endcaps", {
-        // credentials: "include"
-      })
-      .then((response) => response.json())
-      .then((jsonData) => {
-        const endcapData = jsonData.allEndcaps;
-        if(endcapData.length > this.state.endcaps.length) {
-          fetch("http://localhost:5000/settings")
-          .then((response) => response.json())
-          .then((jsonData) => {
-              const currentColNew = jsonData.settings.find((setting) => {
-              return setting.user === this.state.user._id
-            })
-            const currentColumnOrder = currentColNew.columnOrder.endcapIds;
-            const newEndCapId = endcapData[endcapData.length - 1]._id;
-            currentColumnOrder.push(newEndCapId);
-            const settings = {
-              columnOrder: {
-                id: "column-1",
-                title: "To Do",
-                endcapIds: currentColumnOrder,
-              },
-            }
-            // Updates Database with New Order
-              fetch(`http://localhost:5000/settings/${currentColNew._id}`, {
-                method: "PUT",
-                headers: {
-                  "Content-Type": "application/json",
-                },
-                body: JSON.stringify(settings),
-              })
-              .then(() => {
-                const hasUpdated = !this.state.hasUpdated;
-                  // Sets state With New Order
-                  this.setState({
-                    endcaps: endcapData,
-                    columns: {
-                      ...this.state.columns,
-                      "column-1": {
-                        ...this.state.columns["column-1"],
-                        endcapIds: currentColumnOrder,
-                      }
-                    },
-                    hasUpdated: hasUpdated,
-                  })
-              })
-              .catch((err) => console.log(err));
-          })
-        } else {
-          fetch("http://localhost:5000/settings")
-          .then((response) => response.json())
-          .then((jsonData) => {
-            const currentColNew = jsonData.settings.find((setting) => {
-              return setting.user === this.state.user._id
-            })
-            const currentColumnOrder = currentColNew.columnOrder.endcapIds;
-            const hasUpdated = !this.state.hasUpdated;
-            this.setState({
-              endcaps: endcapData,
-              columns: {
-                ...this.state.columns,
-                "column-1": {
-                  ...this.state.columns["column-1"],
-                  endcapIds: currentColumnOrder,
-                }
-              },
-              hasUpdated: hasUpdated,
-            })
-          })
-        }
-      })
-    }
-  })
+  // useEffect(() => {
+  //   handleFetchEndcaps();
+  //   if(updated){
+  //     fetch("http://localhost:5000/endcaps", {
+  //       // credentials: "include"
+  //     })
+  //     .then((response) => response.json())
+  //     .then((jsonData) => {
+  //       const endcapData = jsonData.allEndcaps;
+  //       if(endcapData.length > this.state.endcaps.length) {
+  //         fetch("http://localhost:5000/settings")
+  //         .then((response) => response.json())
+  //         .then((jsonData) => {
+  //             const currentColNew = jsonData.settings.find((setting) => {
+  //             return setting.user === this.state.user._id
+  //           })
+  //           const currentColumnOrder = currentColNew.columnOrder.endcapIds;
+  //           const newEndCapId = endcapData[endcapData.length - 1]._id;
+  //           currentColumnOrder.push(newEndCapId);
+  //           const settings = {
+  //             columnOrder: {
+  //               id: "column-1",
+  //               title: "To Do",
+  //               endcapIds: currentColumnOrder,
+  //             },
+  //           }
+  //           // Updates Database with New Order
+  //             fetch(`http://localhost:5000/settings/${currentColNew._id}`, {
+  //               method: "PUT",
+  //               headers: {
+  //                 "Content-Type": "application/json",
+  //               },
+  //               body: JSON.stringify(settings),
+  //             })
+  //             .then(() => {
+  //               const hasUpdated = !this.state.hasUpdated;
+  //                 // Sets state With New Order
+  //                 this.setState({
+  //                   endcaps: endcapData,
+  //                   columns: {
+  //                     ...this.state.columns,
+  //                     "column-1": {
+  //                       ...this.state.columns["column-1"],
+  //                       endcapIds: currentColumnOrder,
+  //                     }
+  //                   },
+  //                   hasUpdated: hasUpdated,
+  //                 })
+  //             })
+  //             .catch((err) => console.log(err));
+  //         })
+  //       } else {
+  //         fetch("http://localhost:5000/settings")
+  //         .then((response) => response.json())
+  //         .then((jsonData) => {
+  //           const currentColNew = jsonData.settings.find((setting) => {
+  //             return setting.user === this.state.user._id
+  //           })
+  //           const currentColumnOrder = currentColNew.columnOrder.endcapIds;
+  //           const hasUpdated = !this.state.hasUpdated;
+  //           this.setState({
+  //             endcaps: endcapData,
+  //             columns: {
+  //               ...this.state.columns,
+  //               "column-1": {
+  //                 ...this.state.columns["column-1"],
+  //                 endcapIds: currentColumnOrder,
+  //               }
+  //             },
+  //             hasUpdated: hasUpdated,
+  //           })
+  //         })
+  //       }
+  //     })
+  //   }
+  // })
   
   // componentDidMount() {
   //   // if (this.state.isLoggedIn) {
@@ -459,9 +458,10 @@ function App() {
               handleChange={handleChange}
               handleToggleEndcap={handleToggleEndcap}
               handleToggleFlank={handleToggleFlank}
-              columnOrder={columnOrder}
+              // handleFetchEndcaps={handleFetchEndcaps}
+              // columnOrder={columnOrder}
               // columns={this.state.columns}
-              endcaps={endcaps}
+              // endcaps={endcaps}
               onDragEnd={onDragEnd}
               // isLoggedIn={this.state.isLoggedIn}
             />
@@ -491,8 +491,8 @@ function App() {
           <Route path="/edit/:id">
             <EditEndcapPage 
               handleHasUpdated={handleHasUpdated}
-              endcaps={endcaps}
-              columnOrder={columnOrder}
+              // endcaps={endcaps}
+              // columnOrder={columnOrder}
               // user={this.state.user}
             />
           </Route>
